// tools/xxhash-xxh3-64-hash-text-or-file/client/types.ts
type XxHashHashTextOrFileMessages = Readonly<{
  inputLabel: string
  plainTextLabel: string
  plainTextDescription: string
  importFromFileLabel: string
  seedSectionLabel: string
  seedLabel: string
  seedPlaceholder: string
  seedInvalid: string
  hashResultLabel: string
  hashResultDescription: string
  hexLabel: string
  base64Label: string
  decimalLabel: string
  binaryLabel: string
  copyResultLabel: string
  copiedLabel: string
}>

type XxHashHashTextOrFilePageMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
}> &
  XxHashHashTextOrFileMessages

export type { XxHashHashTextOrFileMessages, XxHashHashTextOrFilePageMessages }